extract resource handling into own file
